Edited css selectors and blocks
Tinkered with the declarations as well just to minimize code.
I ended up with a repeat of .para, so I just moved the font-size property and got rid of it.
From .small-para removed font-weight, because it was declared at .para and did not conflict with .para.small-para
From .text .child removed color property because it was declared at the .text block.