Shuffled flexbox properties for .logo ul and .head
I don't understand the point of having the .left- and right-links divs

There is the ul container inside of it that wont inherit the flexbox porperties.

That threw me off when I couldnt put those els in flexbox thorugh using classes.
Then I realized that ul are what need flexboxin.
Only later did I realize that because it was a case of conatiner inside container the flexin needed to be directly applied